+++ compiler/src/analyzer/mod.rs
@@ -1,3 +1,4 @@
+pub mod desugar;
 pub mod monomorphization;
 pub mod type_checker;
 pub mod types;